refactor: extract SpriteCarousel from SpriteDetailDisplay for reusability
